+++ src/EventBus.php
@@ -1,7 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace BSP;
+namespace BSP\EventBus;
+
+use BSP\DrWatson\ExceptionType;
 
 /**
  * EventBus execute listeners listening to a particular event.
@@ -19,12 +21,22 @@ abstract class EventBus
     /** @var EventListener[][] */
     protected $listeners = [];
 
+    /**
+     * @throws EventBusException
+     */
     public function send(Event $event): void
     {
-        if (isset($this->listeners[get_class($event)])) {
-            foreach ($this->listeners[get_class($event)] as $listener) {
-                $listener->listen($event);
-            }
+        if (!isset($this->listeners[get_class($event)])) {
+            throw EventBusException::report(
+                ExceptionType::DOMAIN(),
+                'domain.eventbus.event.unkown',
+                sprintf('"%s" listeners.', static::class),
+                sprintf('You may have forgot to add the "%s" event to the "%s" eventBus.', get_class($event), static::class)
+            );
+        }
+
+        foreach ($this->listeners[get_class($event)] as $listener) {
+            $listener->listen($event);
         }
     }
 }