+++ src/EventBusException.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace BSP\EventBus;
+
+use BSP\DrWatson\DrWatson;
+
+class EventBusException extends DrWatson
+{
+}