+++ tests/Mock/NotifyUserOnUserRegistered.php
@@ -1,10 +1,10 @@
 <?php
 declare(strict_types=1);
 
-namespace BSP\Tests\Mock;
+namespace BSP\EventBus\Tests\Mock;
 
-use BSP\Event;
-use BSP\EventListener;
+use BSP\EventBus\Event;
+use BSP\EventBus\EventListener;
 
 final class NotifyUserOnUserRegistered implements EventListener
 {