+++ CHANGELOG.md
@@ -5,7 +5,11 @@
 
 ### BREAKING CHANGES
 
+- update namespace from `BSP\` to `BSP\EventBus\`
+
 ### New features
+
+-  `\BSP\EventBus` now properly throws exception when event is not declared
 
 ### Bugfixes